Trait Addition and Passion-Capping
Support traits being added to pawns in the UI. Current intent is for locked pawns to be changed in-place, and unlocked pawns to be regenerated as needed.

Support Setting a Cap on Passions. This helps prevent stacking really cracked pawns as much as the UI tended to push me towards.