Refactor Before Introducing AlienRaces-Specific Assembly
I have hit a logistical limit on getting HAR compatibility - the behavior looks perfectly fine for the purposes of single-race scenarios.

However, HAR supports multiple races being in a given scenario, and there's no way to scrounge that data with just reflection magic. So, next steps are to refactor to an actual HAR-specific assembly and have an implementation that can leverage all available information.