Don't run ProcGen on initially opening window
Need to introduce an AllowDirtying property in order to prevent clean state from being dirtied until the first frame of the PrepareProcedurally.cs editor. This way, not even the UI-based incidental edits will force a re-bake of pawns, meaning starting colonists can be kept if users want to use standard randomization for a single pawn, then build around it.