Refactor PawnBuilder.cs into SkillRangeBuilder.cs
Moving the main behavior to a generic abstract implementation allows me to split the SkillDef-based logic out to a more statically-testable form that uses strings instead. With that, I have no dependencies on defs to run tests, and thus nobody needs to set up Rimworld source to be able to attach a debugger.

This allows StringSkillBuilder.cs to essentially have the same behavior as PawnBuilder.cs, but be statically testable and easy-to-debug if I have weird or wrong behavior. This is an improvement over the previous IntegrationTests.cs which I have disabled for now and might just cut out entirely.

Using this, I fixed up some shenanigans that were happening with pawn skill finalization and then wrapped it up with a bow.